pt_dcxt/requirements.txt
Flask==1.0.2
Jinja2==2.10.1
PyYAML==5.1
oyaml==0.5
markdown==2.6.11
